refactor: extract ManagerRegistry + drop sanitizeMetadataForWire re-export (#157 steps 3+5)
Steps 3 and 5 of #157 (after #163, #165) — close out the smaller
mechanical chunks of the KnowledgeBaseServer split.

src/manager-registry.ts (new): RFC 013 M1 per-model FaissIndexManager
cache extracted into its own class. The two private maps and the
init-promise dance that lived on KnowledgeBaseServer are now
ManagerRegistry.getOrCreate(modelId), with the same per-id init
coalescing so concurrent lookups cannot race-construct two managers.

src/KnowledgeBaseServer.ts:
- Replace `managerCache` / `managerInitCache` fields with a single
  `private readonly managers = new ManagerRegistry()`.
- Delete the private `getManagerFor` method; rewrite ~10 internal
  call sites to `this.managers.getOrCreate(...)`. (The wrapper had
  no transport-shape responsibilities — unlike handleKbStats /
  handleListResources which we kept as delegates for the existing
  test surface — so inlining is cleaner.)
- Drop `parseModelId`, `readStoredModelName`, `EmbeddingProvider`
  imports — used only inside the relocated cache.
- Step 5: delete the `export { sanitizeMetadataForWire }` re-export
  (RFC 012 §4.9 boundary fix). Tests in KnowledgeBaseServer.test.ts
  that imported it via `await import('./KnowledgeBaseServer.js')`
  now pull from `./formatter.js` (its canonical home), removing the
  cross-module import bridge.

KnowledgeBaseServer.ts: 795 → 756 (-39). Cumulative across #157
steps 1-3+5: 1052 → 756 (-296). Step 4 (warmup-fanout inversion)
is the only remaining piece and lands separately because it
genuinely changes the host/server contract.

Tests: 527/527 still pass; no new test file (the cache is
exercised transitively by every test that goes through
freshServer(), and the explicit modelName-missing path is unchanged
from the pre-extraction code).

Refs #157.
